requirements.txt
django<4.0.0
django-background-tasks
django-picklefield
linex2
numpy
pandas
matplotlib
networkx
pyvis
tqdm
jinja2
scikit-learn
scipy
statsmodels
natsort
pydantic==1.6.1
regex==2020.9.27
fastapi
jsonschema
gunicorn
psycopg2
typing_extensions
plotly
